Fighting against Git
