+++ src/java/yaw/engine/World.java
@@ -1,20 +1,18 @@
 package yaw.engine;
 
 import org.joml.Quaternionf;
-import org.lwjgl.glfw.GLFWKeyCallback;
 import yaw.engine.camera.Camera;
 import yaw.engine.items.*;
 import yaw.engine.light.SceneLight;
 import yaw.engine.meshs.*;
-import yaw.engine.meshs.strategy.DefaultDrawingStrategy;
 import yaw.engine.skybox.Skybox;
 import org.joml.Vector3f;
 
 import java.util.Vector;
 import java.util.concurrent.ConcurrentHashMap;
 
-import static org.lwjgl.glfw.GLFW.*;
 import static org.lwjgl.opengl.GL11.glViewport;
+import static org.lwjgl.glfw.GLFW.glfwGetTime;
 
 /**
  * Allows the creation of the world and manages through a thread the events, the updates and the rendering to the screen at constant rate.
@@ -26,16 +24,15 @@ public class World implements Runnable {
     private Vector<Camera> mCamerasList;
     private Renderer mRenderer;
     private SceneLight mSceneLight;
-    private KeyCallbackTMP mCallback;
+    private KeyCallback mCallback;
     private Vector<ItemGroup> mItemGroupArrayList;
     private Skybox mSkybox = null;
     private ConcurrentHashMap<String, Texture> mStringTextureConcurrentHashMap;
     private boolean mLoop;
     private int initX, initY, initWidth, initHeight;
     private boolean initVSYNC;
-
+    private WorldNucleus mNucleus;
     private UpdateCallback updateCallback;
-    private KeyCallback keyCallback;
 
     /**
      * Initializes the elements to create the window
@@ -68,19 +65,19 @@ public class World implements Runnable {
     }
 
     public World() {
-        this.mRenderer = new Renderer();
+        this.mNucleus = WorldNucleus.getInstance();
+        this.mRenderer = mNucleus.createRenderer();
         this.mCamerasList = new Vector<>();
-        this.mCamera = new Camera();
-        this.mSceneVertex = new SceneVertex();
-        this.mSceneLight = new SceneLight();
-        this.mCallback = new KeyCallbackTMP();
+        this.mCamera = mNucleus.createCamera();
+        this.mSceneVertex = mNucleus.createSceneVertex();
+        this.mSceneLight = mNucleus.createSceneLight();
+        this.mCallback = mNucleus.createCallback();
         this.mItemGroupArrayList = new Vector<>();
         this.mSkyboxToBeRemoved = new Vector<>();
         this.mLoop = true;
         this.initVSYNC = true;
         this.mStringTextureConcurrentHashMap = new ConcurrentHashMap<>();
-        this.updateCallback = null;
-        this.keyCallback=null;
+        updateCallback = null;
     }
 
     /**
@@ -151,24 +148,23 @@ public class World implements Runnable {
             throw new RuntimeException("RGB must represent 3 colors");
         }
         Vector3f lMaterialColor = new Vector3f(rgb[0], rgb[1], rgb[2]);
-        Material lMaterial = new Material(lMaterialColor);
+        Material lMaterial = mNucleus.createMaterial(lMaterialColor);
         //Texture part
         if (pTextureName != null && !pTextureName.isEmpty()) {
             Texture lTexture = mStringTextureConcurrentHashMap.get(pTextureName);
             if (lTexture == null) {
-                lTexture = new Texture(pTextureName);
+                lTexture = mNucleus.createTexture(pTextureName);
             }
             lMaterial.setTexture(lTexture);
         }
-        Mesh lMesh = new Mesh(pVertices, pTextCoords, pNormals, pIndices, pWeight);
-        lMesh.setDrawingStrategy(new DefaultDrawingStrategy());
+        Mesh lMesh = mNucleus.createMesh(pVertices, pTextCoords, pNormals, pIndices, pWeight);
         lMesh.setMaterial(lMaterial);
         return lMesh;
     }
 
     /**
      * Create a mesh with the specified parameters
-     * MeshOld won't be load into the grahpic cards unless you bind it to an item
+     * Mesh won't be load into the grahpic cards unless you bind it to an item
      *
      * @param pVertices    vetices
      * @param pNormals     normals
@@ -180,9 +176,8 @@ public class World implements Runnable {
             throw new RuntimeException("RGB must represent 3 colors");
         }
         Vector3f lMaterialColor = new Vector3f(rgb[0], rgb[1], rgb[2]);
-        Material lMaterial = new Material(lMaterialColor);
-        Mesh lMesh = new Mesh(pVertices, pNormals, pIndices);
-        lMesh.setDrawingStrategy(new DefaultDrawingStrategy());
+        Material lMaterial = mNucleus.createMaterial(lMaterialColor);
+        Mesh lMesh = mNucleus.createMesh(pVertices, pNormals, pIndices);
         lMesh.setMaterial(lMaterial);
         return lMesh;
     }
@@ -208,15 +203,12 @@ public class World implements Runnable {
 
 
     public boolean isInCollision(HitBox hb1, HitBox hb2) {
-        return hb1.intersect(hb2);
+        return hb1.isIsCollisionWith(hb2);
     }
 
     public synchronized void registerUpdateCallback(UpdateCallback cb) {
         updateCallback = cb;
     }
-
-    public synchronized void registerKeyCallback(KeyCallback key) { keyCallback = key;}
-
 
     /**
      * Allows to initialize the parameters of the class World.
@@ -241,7 +233,7 @@ public class World implements Runnable {
      * @param pB      b
      */
     public void setSkybox(float pWidth, float pLength, float pHeight, float pR, float pG, float pB) {
-        Skybox lSkybox = new Skybox(pWidth, pLength, pHeight, new Vector3f(pR, pG, pB));
+        Skybox lSkybox = mNucleus.createSkybox(pWidth, pLength, pHeight, new Vector3f(pR, pG, pB));
         setSkybox(lSkybox);
     }
 
@@ -305,10 +297,6 @@ public class World implements Runnable {
     private void UFRV() throws InterruptedException {
         double dt = 0.01; // Update Rate: 1 ~= 2 fps | 0.001 ~= 1000 fps
         /* Initialization of the window we currently use. */
-        int key;
-        int scancode;
-        int action;
-        int mods;
         glViewport(initX, initY, initWidth, initHeight);
         double beforeTime = glfwGetTime();
         double lag = 0d;
@@ -329,29 +317,6 @@ public class World implements Runnable {
                     lag -= dt;
                 }
             }
-
-
-            // TODO : is this going in the right direction ? is the idea to get the key pressed here and send its information to the KeyCallBack the right idea ?
-
-            if(Input.isKeyDown(GLFW_KEY_UP)){
-                key=GLFW_KEY_UP;
-                scancode= glfwGetKeyScancode(key);
-                action =GLFW_PRESS;
-                mods = 0;
-            }else{
-                key=GLFW_KEY_DOWN;
-                scancode= glfwGetKeyScancode(key);
-                action =GLFW_PRESS;
-                mods = 0;
-            }
-
-
-
-            if(keyCallback!= null){
-                keyCallback.sendKey(key, scancode,action,mods);
-            }
-
-
 
             /*Clean the window*/
             boolean isResized = Window.clear();
@@ -475,7 +440,7 @@ public class World implements Runnable {
         return mSceneLight;
     }
 
-    public KeyCallbackTMP getCallback() {
+    public KeyCallback getCallback() {
         return mCallback;
     }
 }
