+++ src/java/yaw/engine/items/Item.java
@@ -97,19 +97,28 @@ public abstract class Item {
      * Rotate along X axis
      * @param angle of rotation (in degree)
      */
-    public abstract void rotateX(float angle);
+    public void rotateX(float angle) {
+        orientation.rotateX(toRadians(angle));
+        invalidate();
+    }
 
     /**
      * Rotate along Y axis
      * @param angle of rotation (in degree)
      */
-    public abstract void rotateY(float angle);
+    public void rotateY(float angle) {
+        orientation.rotateY(toRadians(angle));
+        invalidate();
+    }
 
     /**
      * Rotate along Z axis
      * @param angle of rotation (in degree)
      */
-    public abstract void rotateZ(float angle);
+    public void rotateZ(float angle) {
+        orientation.rotateZ(toRadians(angle));
+        invalidate();
+    }
 
     /**
      * Rotation along three axes (Euler angles rotation)
@@ -117,14 +126,20 @@ public abstract class Item {
      * @param angleY same for axis Y
      * @param angleZ same for axis Z
      */
-    public abstract void rotateXYZ(float angleX, float angleY, float angleZ) ;
+    public void rotateXYZ(float angleX, float angleY, float angleZ) {
+        orientation.rotateXYZ(toRadians(angleX), toRadians(angleY), toRadians(angleZ));
+        invalidate();
+    }
 
     /**
      * Rotation of given angle along an axis
      * @param angle angle of rotation in degrees
      * @param axis axis of rotation
      */
-    public abstract void rotateAxis(float angle, Vector3f axis) ;
+    public void rotateAxis(float angle, Vector3f axis) {
+        orientation.rotateAxis(toRadians(angle), axis);
+        invalidate();
+    }
 
     /**
      * Rotation of given angle along an axis, and aroung the specified center
@@ -132,28 +147,52 @@ public abstract class Item {
      * @param axis the axis of rotation
      * @param center the center of rotation
      */
-    public abstract void rotateAxisAround(float angle, Vector3f axis, Vector3f center);
+    public void rotateAxisAround(float angle, Vector3f axis, Vector3f center) {
+        axis = axis.normalize();
+        // change orientation
+        orientation.rotateAxis(toRadians(angle), axis);
+
+        // change position
+        new Matrix4f().rotateAround(new Quaternionf(new AxisAngle4f(toRadians(angle), axis)), center.x, center.y, center.z)
+                .transformPosition(position);
+
+/*
+        new Matrix4f()
+                .translate(center)
+                .rotate(toRadians(angle), axis)
+                .translate(center.negate())
+                .transformPosition(position);
+*/
+
+        invalidate();
+    }
 
     /**
      * Rotate along X axis, around center
      * @param angle of rotation (in degree)
      * @param center the center of rotation
      */
-    public  abstract void rotateXAround(float angle, Vector3f center) ;
+    public final void rotateXAround(float angle, Vector3f center) {
+        rotateAxisAround(angle, new Vector3f(1.0f, 0.0f, 0.0f), center);
+    }
 
     /**
      * Rotate along Y axis, around center
      * @param angle of rotation (in degree)
      * @param center the center of rotation
      */
-    public abstract void rotateYAround(float angle, Vector3f center) ;
+    public final void rotateYAround(float angle, Vector3f center) {
+        rotateAxisAround(angle, new Vector3f(0.0f, 1.0f, 0.0f), center);
+    }
 
     /**
      * Rotate along Z axis, around center
      * @param angle of rotation (in degree)
      * @param center the center of rotation
      */
-    public abstract void rotateZAround(float angle, Vector3f center);
+    public final void rotateZAround(float angle, Vector3f center) {
+        rotateAxisAround(angle, new Vector3f(0.0f, 0.0f, 1.0f), center);
+    }
 
     /**
      * Rotation along three axes (Euler angles rotation), around center
@@ -162,7 +201,12 @@ public abstract class Item {
      * @param angleZ same for axis Z
      * @param center the center of rotation
      */
-    public abstract void rotateXYZAround(float angleX, float angleY, float angleZ, Vector3f center) ;
+    public final void rotateXYZAround(float angleX, float angleY, float angleZ, Vector3f center) {
+        AxisAngle4f aaxis = new AxisAngle4f(new Quaternionf().rotationXYZ(toRadians(angleX)
+                                                    , toRadians(angleY)
+                                                    , toRadians(angleZ))).normalize();
+        rotateAxisAround(toDegrees(aaxis.angle), new Vector3f(aaxis.x, aaxis.y, aaxis.z), center);
+    }
 
 
     // XXX : use this ?
