+++ src/java/yaw/engine/items/ItemGroup.java
@@ -162,48 +162,6 @@ public class ItemGroup extends Item {
         positionAtCentroid();
     }
 
-    /**
-     * Rotation along three axes (Euler angles rotation), around center
-     * @param angleX angle of rotation along axis X (in degrees)
-     * @param angleY same for axis Y
-     * @param angleZ same for axis Z
-     * @param center the center of rotation
-     */
-    public void rotateXYZAround(float angleX, float angleY, float angleZ, Vector3f center) {
-        AxisAngle4f aaxis = new AxisAngle4f(new Quaternionf().rotationXYZ(toRadians(angleX)
-                , toRadians(angleY)
-                , toRadians(angleZ))).normalize();
-        rotateAxisAround(toDegrees(aaxis.angle), new Vector3f(aaxis.x, aaxis.y, aaxis.z), center);
-    }
-
-
-    /**
-     * Rotate along X axis, around center
-     * @param angle of rotation (in degree)
-     * @param center the center of rotation
-     */
-    public void rotateXAround(float angle, Vector3f center) {
-        rotateAxisAround(angle, new Vector3f(1.0f, 0.0f, 0.0f), center);
-    }
-
-    /**
-     * Rotate along Y axis, around center
-     * @param angle of rotation (in degree)
-     * @param center the center of rotation
-     */
-    public void rotateYAround(float angle, Vector3f center) {
-        rotateAxisAround(angle, new Vector3f(0.0f, 1.0f, 0.0f), center);
-    }
-
-    /**
-     * Rotate along Z axis, around center
-     * @param angle of rotation (in degree)
-     * @param center the center of rotation
-     */
-    public void rotateZAround(float angle, Vector3f center) {
-        rotateAxisAround(angle, new Vector3f(0.0f, 0.0f, 1.0f), center);
-    }
-
     public void removeAll() {
         items.clear();
         position=new Vector3f(0f, 0f, 0f);
