+++ src/java/yaw/engine/items/ItemObject.java
@@ -1,6 +1,9 @@
 package yaw.engine.items;
 
-import org.joml.*;
+import org.joml.Matrix4f;
+import org.joml.Quaternionf;
+import org.joml.Vector3f;
+import org.joml.Vector4f;
 import yaw.engine.meshs.Material;
 import yaw.engine.meshs.Mesh;
 
@@ -56,123 +59,6 @@ public class ItemObject extends Item {
      * Get the mesh (geometry) of the item
      */
     public Mesh getMesh(){ return mesh; }
-
-    /**
-     * Rotate along X axis
-     * @param angle of rotation (in degree)
-     */
-    public void rotateX(float angle) {
-        orientation.rotateX(toRadians(angle));
-        invalidate();
-    }
-
-    /**
-     * Rotate along Y axis
-     * @param angle of rotation (in degree)
-     */
-    public void rotateY(float angle) {
-        orientation.rotateY(toRadians(angle));
-        invalidate();
-    }
-
-    /**
-     * Rotate along Z axis
-     * @param angle of rotation (in degree)
-     */
-    public void rotateZ(float angle) {
-        orientation.rotateZ(toRadians(angle));
-        invalidate();
-    }
-
-    /**
-     * Rotation along three axes (Euler angles rotation)
-     * @param angleX angle of rotation along axis X (in degrees)
-     * @param angleY same for axis Y
-     * @param angleZ same for axis Z
-     */
-    public void rotateXYZ(float angleX, float angleY, float angleZ) {
-        orientation.rotateXYZ(toRadians(angleX), toRadians(angleY), toRadians(angleZ));
-        invalidate();
-    }
-
-    /**
-     * Rotation of given angle along an axis
-     * @param angle angle of rotation in degrees
-     * @param axis axis of rotation
-     */
-    public void rotateAxis(float angle, Vector3f axis) {
-        orientation.rotateAxis(toRadians(angle), axis);
-        invalidate();
-    }
-
-    /**
-     * Rotation of given angle along an axis, and aroung the specified center
-     * @param angle the angle of rotation (in degrees)
-     * @param axis the axis of rotation
-     * @param center the center of rotation
-     */
-    public void rotateAxisAround(float angle, Vector3f axis, Vector3f center) {
-        axis = axis.normalize();
-        // change orientation
-        orientation.rotateAxis(toRadians(angle), axis);
-
-        // change position
-        new Matrix4f().rotateAround(new Quaternionf(new AxisAngle4f(toRadians(angle), axis)), center.x, center.y, center.z)
-                .transformPosition(position);
-
-//        new Matrix4f()
-//                .translate(center)
-//                .rotate(toRadians(angle), axis)
-//                .translate(new Vector3f(center).negate())
-//                .transformPosition(position);
-
-        invalidate();
-    }
-
-    /**
-     * Rotate along X axis, around center
-     * @param angle of rotation (in degree)
-     * @param center the center of rotation
-     */
-    public void rotateXAround(float angle, Vector3f center) {
-        rotateAxisAround(angle, new Vector3f(1.0f, 0.0f, 0.0f), center);
-    }
-
-    /**
-     * Rotate along Y axis, around center
-     * @param angle of rotation (in degree)
-     * @param center the center of rotation
-     */
-    public void rotateYAround(float angle, Vector3f center) {
-        rotateAxisAround(angle, new Vector3f(0.0f, 1.0f, 0.0f), center);
-    }
-
-    /**
-     * Rotate along Z axis, around center
-     * @param angle of rotation (in degree)
-     * @param center the center of rotation
-     */
-    public void rotateZAround(float angle, Vector3f center) {
-        rotateAxisAround(angle, new Vector3f(0.0f, 0.0f, 1.0f), center);
-    }
-
-    /**
-     * Rotation along three axes (Euler angles rotation), around center
-     * @param angleX angle of rotation along axis X (in degrees)
-     * @param angleY same for axis Y
-     * @param angleZ same for axis Z
-     * @param center the center of rotation
-     */
-    public void rotateXYZAround(float angleX, float angleY, float angleZ, Vector3f center) {
-        AxisAngle4f aaxis = new AxisAngle4f(new Quaternionf().rotationXYZ(toRadians(angleX)
-                , toRadians(angleY)
-                , toRadians(angleZ))).normalize();
-        rotateAxisAround(toDegrees(aaxis.angle), new Vector3f(aaxis.x, aaxis.y, aaxis.z), center);
-    }
-
-
-
-
